feat: Seperate app.js and server.js
